+++ src/Components/Nav.js
@@ -1,6 +1,7 @@
 import React from 'react';
 
 import './Nav.scss';
+
 const Nav = () => {
   return (
     <section className="mainNav">
