+++ src/pages/hyeongkyeom/Components/Story.js
@@ -1,0 +1,114 @@
+import React from 'react';
+
+const Story = () => {
+  return (
+    <div className="mainStory">
+      <section className="storyBorder">
+        <div className="storyProfile">
+          <div className="storyProfileOne">
+            <img
+              className="storyImg"
+              src="images/hyeongkyeom/profile/pro1.jpeg"
+              alt="storyImg1"
+            />
+            <div className="storyProfileName">
+              <span>all.about.seoul.trip</span>
+            </div>
+          </div>
+          <div className="storyProfileOne">
+            <img
+              className="storyImg"
+              src="images/hyeongkyeom/profile/pro2.jpeg"
+              alt="storyImg2"
+            />
+            <div className="storyProfileName">
+              <span>kangnam11</span>
+            </div>
+          </div>
+          <div className="storyProfileOne">
+            <img
+              className="storyImg"
+              src="images/hyeongkyeom/profile/pro3.jpeg"
+              alt="storyImg3"
+            />
+            <div className="storyProfileName">
+              <span>nyamnyamang</span>
+            </div>
+          </div>
+          <div className="storyProfileOne">
+            <img
+              className="storyImg"
+              src="images/hyeongkyeom/profile/pro6.jpeg"
+              alt="storyImg4"
+            />
+            <div className="storyProfileName">
+              <span>all.about.20s</span>
+            </div>
+          </div>
+          <div className="storyProfileOne">
+            <img
+              className="storyImg"
+              src="images/hyeongkyeom/profile/pro4.jpeg"
+              alt="storyImg5"
+            />
+            <div className="storyProfileName">
+              <span>seoultravel</span>
+            </div>
+          </div>
+          <div className="storyProfileOne">
+            <img
+              className="storyImg"
+              src="images/hyeongkyeom/profile/pro5.jpeg"
+              alt="storyImg6"
+            />
+            <div className="storyProfileName">
+              <span>datecourse123</span>
+            </div>
+          </div>
+          <div className="storyProfileOne">
+            <img
+              className="storyImg"
+              src="images/hyeongkyeom/profile/pro6.jpeg"
+              alt="storyImg7"
+            />
+            <div className="storyProfileName">
+              <span>leagueoflegend</span>
+            </div>
+          </div>
+          <div className="storyProfileOne">
+            <img
+              className="storyImg"
+              src="images/hyeongkyeom/profile/pro1.jpeg"
+              alt="storyImg8"
+            />
+            <div className="storyProfileName">
+              <span>love_doong.doong</span>
+            </div>
+          </div>
+          <div className="storyProfileOne">
+            <img
+              className="storyImg"
+              src="images/hyeongkyeom/profile/pro2.jpeg"
+              alt="storyImg9"
+            />
+            <div className="storyProfileName">
+              <span>marvelkorea</span>
+            </div>
+          </div>
+          <div className="storyProfileOne">
+            <img
+              className="storyImg"
+              src="images/hyeongkyeom/profile/pro3.jpeg"
+              alt="storyImg10"
+            />
+            <div className="storyProfileName">
+              <span>cheongjeul</span>
+            </div>
+          </div>
+        </div>
+      </section>
+    </div>
+  );
+};
+
+export default Story;
