+++ src/pages/hyeongkyeom/Main/Main.scss
@@ -31,7 +31,7 @@ body {
       padding-left: 10px;
       font-size: 20px;
 
-      :hover {
+      &:hover {
         background-color: #fafafa;
       }
     }
@@ -65,7 +65,7 @@ body {
     border: 1px solid #efefef;
     border-radius: 5px;
 
-    :after {
+    &:after {
       border-top: 0px solid transparent;
       border-left: 10px solid transparent;
       border-right: 10px solid transparent;
@@ -123,8 +123,9 @@ body {
 
   .feeds {
     margin: 70px auto;
+    margin-right: 23rem;
     width: 616px;
-    background-color: white;
+    background-color: #fafafa;
   }
 
   .feedProfile {
@@ -189,8 +190,10 @@ body {
     border-top: none;
     border-radius: 0 0 5px 5px;
     display: flex;
+    height: 100px;
     flex-direction: row;
     height: 50px;
+    margin-bottom: 20px;
   }
 
   .addComment {
@@ -199,7 +202,7 @@ body {
     color: #3b99ef;
     cursor: pointer;
 
-    :disabled {
+    &:disabled {
       color: #8e8e8e;
       cursor: auto;
     }
@@ -290,15 +293,15 @@ body {
     opacity: 0;
     transition: 0.3s;
 
-    :hover {
+    &:hover {
       opacity: 1;
     }
   }
 
   .mainRight {
     position: fixed;
-    right: 150px;
-    top: 110px;
+    right: 50px;
+    top: 100px;
     width: 350px;
   }
 
@@ -336,6 +339,7 @@ body {
   .mainStory {
     display: flex;
     background-color: #fafafa;
+    margin: 15px auto;
   }
 
   .storyBorder {
